+++ src/Http/Controllers/ManifestController.php
@@ -6,13 +6,13 @@ namespace SlFomin\PwaLaravel\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
-use SlFomin\PwaLaravel\Contracts\ManifestDriver;
+use SlFomin\PwaLaravel\PwaManager;
 
 final class ManifestController
 {
-    public function __invoke(Request $request, ManifestDriver $driver): Response
+    public function __invoke(Request $request, PwaManager $pwa): Response
     {
-        $manifest = $driver->resolve($request);
+        $manifest = $pwa->manifest($request);
 
         return response(
             $manifest->toJson(),